+++ subsystems/posix/sdk/v1/runtime/posix.c
@@ -846,6 +846,7 @@ char* sx_strerror(int error_number) {
         case ERANGE: return "result out of range";
         case ENOSYS: return "function not implemented";
         case ENOTEMPTY: return "directory not empty";
+        case E2BIG: return "argument list too long";
         case ETIMEDOUT: return "timed out";
         default: return "error";
     }